Documentation and STATIC plugin?

+++ plugins/BlockCompressor/BlockCompressor.cpp
@@ -63,10 +63,10 @@ BlockCompressor::~BlockCompressor()
 
 void BlockCompressor::write_elias_fano()
 {
-    std::size_t ef_size = ef_pos.size();
+    std::uint64_t ef_size = ef_pos.size();
 
     ef_out.write(reinterpret_cast<const char*>(&minimum_hash), sizeof(std::uint64_t));
-    ef_out.write(reinterpret_cast<const char*>(&ef_size), sizeof(std::size_t));
+    ef_out.write(reinterpret_cast<const char*>(&ef_size), sizeof(std::uint64_t));
     
     //Create Elias-Fano representation from positions
     sdsl::sd_vector<> ef(ef_pos.begin(), ef_pos.end());
